+++ src/rules/date.js
@@ -1,6 +1,5 @@
 const moment = require('moment');
 
-
 module.exports = async function date(field, value, format = 'YYYY-MM-DD') {
 
     if (!moment(value, format, true).isValid()) {