+++ src/rules/dateiso.js
@@ -1,0 +1,12 @@
+const moment = require('moment');
+
+module.exports = async function date(field, value) {
+
+    if (!moment(value, moment.ISO_8601).isValid()) {
+
+        return false;
+    }
+
+    return true;
+
+}