+++ src/rules/dimensions.js
@@ -1,0 +1,19 @@
+const sizeOf = require('image-size');
+
+module.exports = async function dimensions(field, file, args) {
+
+    let max, min;
+
+    if (!Array.isArray(args)) {
+        args = [args];
+    }
+
+    max = args[0];
+    if (args.length >= 2) {
+        min = args[1];
+    }
+
+    const dimensions = sizeOf(file.path);
+
+    //dimensions.width, dimensions.height
+}